requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

numpy==1.23.5
opencv-python==4.7.0.68
streamlit==1.17.0
streamlit-drawable-canvas==0.9.2
tensorflow-intel==2.11.0
